Turned calculateReturnValue method into a double method; Added a retailPrice field to the Product class; Starting creation of Store methods

+++ src/main/java/org/example/GeneralProduct.java
@@ -20,7 +20,7 @@ public class GeneralProduct extends Product implements Returnable {
     }
 
     @Override
-    public int calculateReturnValue() {
+    public double calculateReturnValue() {
         return (isUsed) ? 0 : getPrice();
     }
